requirements.txt
alembic==1.6.5
asgiref==3.3.4
Babel==2.9.1
bcrypt==3.2.0
blinker==1.4
Brotli==1.0.9
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cheroot==8.5.2
click==7.1.2
cryptography==3.4.7
dateparser==1.0.0
defusedxml==0.7.1
Django==3.2.3
django-3-jet==1.0.7
django-admin-autocomplete-filter==0.6.1
django-admin-list-filter-dropdown==1.0.3
django-ajax-datatable==4.2.0
django-autocompletefilter==0.0.7
django-environ==0.8.1
django-ip==1.0.2
django-six==1.0.4
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.4
Flask-BabelEx==0.9.4
Flask-Compress==1.9.0
Flask-Gravatar==0.5.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security-Too==3.4.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
google-api-python-client==1.4.1
gunicorn==20.1.0
httpagentparser==1.9.1
httplib2==0.19.1
idna==2.10
itsdangerous==1.1.0
jaraco.functools==3.3.0
jet-bridge-base==1.0.1
Jinja2==2.11.3
ldap3==2.9
Mako==1.1.4
MarkupSafe==2.0.1
more-itertools==8.8.0
oauth2client==4.1.3
oauthlib==3.1.1
paramiko==2.7.2
passlib==1.7.4
pgadmin4==5.3
Pillow==8.2.0
prompt-toolkit==2.0.9
psutil==5.8.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
python3-openid==3.2.0
pytz==2021.1
razorpay==1.2.0
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
simplejson==3.17.2
six==1.16.0
social-auth-core==4.1.0
speaklater==1.3
speaklater3==1.4
SQLAlchemy==1.3.24
sqlparse==0.4.1
sshtunnel==0.4.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.5
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
